refactor: split monolith constants file into domain-specific modules
