+++ src/apps/content/components/mobile/MobileFab.vue
@@ -46,7 +46,8 @@ import { useMobileStore } from '@/store/modules/mobile.js';
 import { storageManager } from '@/shared/storage/core/StorageCore.js';
 import { useUnifiedI18n } from '@/composables/shared/useUnifiedI18n.js';
 import { useResourceTracker } from '@/composables/core/useResourceTracker.js';
-import { MOBILE_CONSTANTS, TRANSLATION_STATUS } from '@/shared/config/constants.js';
+import { MOBILE_CONSTANTS } from '@/shared/constants/mobile.js';
+import { TRANSLATION_STATUS } from '@/shared/constants/translation.js';
 import { deviceDetector } from '@/utils/browser/compatibility.js';
 import { getScopedLogger } from '@/shared/logging/logger.js';
 import { LOG_COMPONENTS } from '@/shared/logging/logConstants.js';
