+++ src/features/element-selection/SelectElementManager.js
@@ -10,7 +10,8 @@ import { MessageActions } from '@/shared/messaging/core/MessageActions.js';
 import ExtensionContextManager from '@/core/extensionContext.js';
 import { isFatalError, isCancellationError } from '@/shared/error-management/ErrorMatcher.js';
 import { getSettingsAsync } from '@/shared/config/config.js';
-import { NOTIFICATION_TIME, TRANSLATION_STATUS } from '@/shared/config/constants.js';
+import { NOTIFICATION_TIME } from '@/shared/constants/ui.js';
+import { TRANSLATION_STATUS } from '@/shared/constants/translation.js';
 import { getTranslationString } from '@/utils/i18n/i18n.js';
 import { shouldShowProviderWarning } from '@/shared/utils/warning-manager.js';
 import { ProviderRegistryIds } from '@/features/translation/providers/ProviderConstants.js';
