+++ src/features/translation/providers/LanguageSwappingService.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { LanguageSwappingService } from './LanguageSwappingService.js';
-import { AUTO_DETECT_VALUE } from "@/shared/config/constants.js";
+import { AUTO_DETECT_VALUE } from "@/shared/constants/core.js";
 
 // Mock dependencies
 vi.mock('@/shared/logging/logger.js', () => ({
